requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
alpha-transform==0.0.1
anyio==3.5.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
astra-toolbox==2.1.2
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.4.0
babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blessed==1.19.1
blinker==1.4
bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.0.0
cryptography==37.0.1
cycler==0.11.0
cytoolz==0.11.0
dask==2021.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2022.7.1
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.4
google-auth==2.6.0
gpustat==1.0.0
grpcio==1.42.0
h5py==3.7.0
healpy==1.16.1
idna==3.4
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.2.0
ipykernel==6.15.2
ipython-genutils==0.2.0
ipython==7.31.1
ipywidgets==7.6.5
jedi==0.18.1
jinja2==3.1.2
joblib==1.1.1
json5==0.9.6
jsonschema==4.16.0
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.23.4
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.16.5
jupyterlab-widgets==1.0.0
jupyterlab==3.5.2
kiwisolver==1.4.2
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.1
markdown==3.3.4
markupsafe==2.1.1
matplotlib-inline==0.1.6
matplotlib==3.5.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook-shim==0.2.2
notebook==6.5.2
numba==0.56.3
numexpr==2.8.3
numpy==1.21.5
nvidia-ml-py==11.495.46
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
pip==22.2.2
pkgutil-resolve-name==1.3.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyfftw==0.13.0
pygments==2.11.2
pyjwt==2.4.0
pyopenssl==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywavelets==1.3.0
pyyaml==6.0
pyzmq==23.2.0
qtconsole==5.4.0
qtpy==2.2.0
requests-oauthlib==1.3.0
requests==2.28.1
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
send2trash==1.8.0
sentry-sdk==1.9.9
setproctitle==1.3.2
setuptools==63.4.1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.0
terminado==0.17.1
threadpoolctl==2.2.0
tifffile==2020.10.1
timm==0.6.7
tinycss2==1.2.1
tomli==2.0.1
toolz==0.11.2
torch-radon==1.0.0
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing-extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
yarl==1.8.1
zipp==3.8.1